[Gemini] Refactor memstats
