+++ colossalai/gemini/memory_tracer/__init__.py
@@ -1,5 +1,10 @@
-from .model_data_memtracer import GLOBAL_MODEL_DATA_TRACER
+from .chunk_memstats_collector import ChunkMemStatsCollector
 from .memory_monitor import AsyncMemoryMonitor, SyncCudaMemoryMonitor
 from .memstats_collector import MemStatsCollector
+from .model_data_memtracer import GLOBAL_MODEL_DATA_TRACER
+from .static_memstats_collector import StaticMemStatsCollector
 
-__all__ = ['AsyncMemoryMonitor', 'SyncCudaMemoryMonitor', 'MemStatsCollector', 'GLOBAL_MODEL_DATA_TRACER']
+__all__ = [
+    'AsyncMemoryMonitor', 'SyncCudaMemoryMonitor', 'MemStatsCollector', 'ChunkMemStatsCollector',
+    'StaticMemStatsCollector', 'GLOBAL_MODEL_DATA_TRACER'
+]
